misc clean-up

+++ unit_tests/engine_test_helper.cpp
@@ -181,7 +181,7 @@ void EngineTestHelper::assertEvent(const char *msg, int index, void *callback, e
 
 	InjectionEvent *actualEvent = (InjectionEvent *)event->action.getArgument();
 
-	assertEqualsLM(msg, expectedEvent->outputs[0], (long)actualEvent->outputs[0]);
+	assertEqualsLM(msg, (long)expectedEvent->outputs[0], (long)actualEvent->outputs[0]);
 // but this would not work	assertEqualsLM(msg, expectedPair, (long)eventPair);
 }
